Extract server logic
