+++ boundary/main.py
@@ -1,11 +1,9 @@
 from argparse import ArgumentParser
 from configparser import ConfigParser
 import os
-import socket
 import logging
+from server import Server
 
-MAX_READ_SIZE = 1024
-LENGTH_BYTES = 2
 
 def initialize_log(logging_level):
     """
@@ -58,51 +56,12 @@ def parse_args() -> list[str]:
     return parser.parse_args().destination_queue
 
 
-def recieve_exact(s: socket.socket, length: int) -> bytes:
-    data = b''
-    while len(data) < length:
-        bytes_remaining = length - len(data)
-        new_data = s.recv(min(MAX_READ_SIZE, bytes_remaining))
-        if not new_data:
-            raise ConnectionResetError("Connection closed")
-        data += new_data
-    return data
-
-
-def receive_line(s: socket.socket) -> bytes:
-    length_bytes = recieve_exact(s, LENGTH_BYTES)
-    length = int.from_bytes(length_bytes, byteorder='big')
-    data = recieve_exact(s, length)
-    return data
-
-
-def handle_client_connection(client_socket: socket.socket):
-    while True:
-        try:
-            data = receive_line(client_socket)
-            logging.debug("Received line: %s", data.decode().strip())
-            # TODO: Send the data to RabbitMQ
-        except ConnectionResetError:
-            logging.info("Connection closed by client")
-            break
-
-
 def main():
     config_params = initialize_config()
     initialize_log(config_params["logging_level"])
     destination_queue = parse_args()
-
-    server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    server_socket.bind(('', config_params["port"]))
-    server_socket.listen(config_params["listen_backlog"])
-    logging.info(
-        "Listening for connections and redirecting to %s", destination_queue)
-
-    while True:
-        client_socket, address = server_socket.accept()
-        logging.info("Connection from %s", address)
-        with client_socket:
-            handle_client_connection(client_socket)
+    server = Server(config_params["port"], config_params["listen_backlog"], destination_queue)
+    server.run()
 
 
 if __name__ == "__main__":
